refactor: move voice failure tracking to media limits
